modified fetchProduct function for components
